Sto/requirements.txt
google-cloud-storage == 2.7.0